fix: improve bug reporting url
