+++ src/alva-util/index.ts
@@ -1,6 +1,7 @@
 export * from './compute-difference';
 export * from './drag-and-drop';
 export * from './ensure-array';
+export * from './new-issue-url';
 export * from './guess-name';
 export * from './noop';
 export * from './parse-json';
